refactor(settings): extract PackingListsPage
